+++ src/util.ts
@@ -1,30 +1,5 @@
 import { RemoteBranch } from './types'
 
-export function normalizedExtension(ext: string): string {
-  if (!ext) {
-    return ext
-  }
-  if (ext.indexOf('.') === 0) {
-    return ext
-  }
-  return `.${ext}`
-}
-
-export function parseExtensions(input?: string): string[] {
-  if (!input) return []
-
-  return input
-    .trim()
-    .split(' ')
-    .map(e => {
-      return e.trim()
-    })
-    .filter(e => {
-      return e.length > 0
-    })
-    .map(normalizedExtension)
-}
-
 export function remoteBranchString(param: RemoteBranch): string {
   const { remote, branch } = param
   const prefix = `${remote}/`